+++ src/v1/Kucoin.php
@@ -7,6 +7,7 @@ namespace Feedex\Kucoin\v1;
 use Feedex\Contracts\Capabilities\HasAccountModuleInterface;
 use Feedex\Contracts\Capabilities\HasAssetSpotBalanceModuleInterface;
 use Feedex\Contracts\Capabilities\HasCommonModuleInterface;
+use Feedex\Contracts\Capabilities\HasFuturesMarketCoreModuleInterface;
 use Feedex\Contracts\Capabilities\HasSpotDealModuleInterface;
 use Feedex\Contracts\Capabilities\HasSpotMarketCoreModuleInterface;
 use Feedex\Contracts\Capabilities\HasSpotOrderCoreModuleInterface;
@@ -15,6 +16,7 @@ use Feedex\Kucoin\v1\Http\KucoinHttpClient;
 use Feedex\Kucoin\v1\Modules\Account;
 use Feedex\Kucoin\v1\Modules\Asset;
 use Feedex\Kucoin\v1\Modules\Common;
+use Feedex\Kucoin\v1\Modules\FuturesMarket;
 use Feedex\Kucoin\v1\Modules\SpotDeal;
 use Feedex\Kucoin\v1\Modules\SpotMarket;
 use Feedex\Kucoin\v1\Modules\SpotOrder;
@@ -26,18 +28,22 @@ final class Kucoin implements
     HasAssetSpotBalanceModuleInterface,
     HasSpotMarketCoreModuleInterface,
     HasSpotOrderCoreModuleInterface,
-    HasSpotDealModuleInterface
+    HasSpotDealModuleInterface,
+    HasFuturesMarketCoreModuleInterface
 {
-    private KucoinHttpClient $httpClient;
+    private KucoinHttpClient $spotHttpClient;
+    private KucoinHttpClient $futuresHttpClient;
 
     public function __construct(
         string $apiKey,
         string $apiSecret,
         string $apiPassphrase,
         string $baseUrl = 'https://api.kucoin.com',
+        string $futuresBaseUrl = 'https://api-futures.kucoin.com',
         int $timeout = 60
     ) {
-        $this->httpClient = new KucoinHttpClient($apiKey, $apiSecret, $apiPassphrase, $baseUrl, $timeout);
+        $this->spotHttpClient = new KucoinHttpClient($apiKey, $apiSecret, $apiPassphrase, $baseUrl, $timeout);
+        $this->futuresHttpClient = new KucoinHttpClient($apiKey, $apiSecret, $apiPassphrase, $futuresBaseUrl, $timeout);
     }
 
     public function id(): string
@@ -47,31 +53,36 @@ final class Kucoin implements
 
     public function common(): Common
     {
-        return new Common($this->httpClient);
+        return new Common($this->spotHttpClient);
     }
 
     public function account(): Account
     {
-        return new Account($this->httpClient);
+        return new Account($this->spotHttpClient);
     }
 
     public function asset(): Asset
     {
-        return new Asset($this->httpClient);
+        return new Asset($this->spotHttpClient);
     }
 
     public function spotMarket(): SpotMarket
     {
-        return new SpotMarket($this->httpClient);
+        return new SpotMarket($this->spotHttpClient);
     }
 
     public function spotOrder(): SpotOrder
     {
-        return new SpotOrder($this->httpClient);
+        return new SpotOrder($this->spotHttpClient);
     }
 
     public function spotDeal(): SpotDeal
     {
-        return new SpotDeal($this->httpClient);
+        return new SpotDeal($this->spotHttpClient);
+    }
+
+    public function futuresMarket(): FuturesMarket
+    {
+        return new FuturesMarket($this->futuresHttpClient);
     }
 }