+++ tests/KucoinTest.php
@@ -8,6 +8,7 @@ use Feedex\Kucoin\v1\Kucoin;
 use Feedex\Kucoin\v1\Modules\Account;
 use Feedex\Kucoin\v1\Modules\Asset;
 use Feedex\Kucoin\v1\Modules\Common;
+use Feedex\Kucoin\v1\Modules\FuturesMarket;
 use Feedex\Kucoin\v1\Modules\SpotDeal;
 use Feedex\Kucoin\v1\Modules\SpotMarket;
 use Feedex\Kucoin\v1\Modules\SpotOrder;
@@ -26,5 +27,6 @@ final class KucoinTest extends TestCase
         self::assertInstanceOf(SpotMarket::class, $client->spotMarket());
         self::assertInstanceOf(SpotOrder::class, $client->spotOrder());
         self::assertInstanceOf(SpotDeal::class, $client->spotDeal());
+        self::assertInstanceOf(FuturesMarket::class, $client->futuresMarket());
     }
 }